refactor: merge specs
